Fixing cryptography dependency pinning
For some reason the folks at cryptography decided to flip the table
on their versioning schema and went from 3.4.8 to 35.0.

This is now causing build failures for the dbt-snowflake package, as
it's pinning the max version to be `<4`, while other dependencies of
this package don't have such restriction, causing a conflict within
available and required versions.